requirements.txt
# Deep Learning - GPU
tensorflow==2.15.0

# Data
pandas
scikit-learn==1.8.0


# Images
Pillow==10.2.0
opencv-python==4.9.0.80
albumentations==1.4.3

# Visualisation
matplotlib==3.8.3
seaborn==0.13.2
tf-keras-vis==0.8.5

# Utilitaires
tqdm==4.66.2
python-dotenv==1.0.1
pyyaml==6.0.1
split-folders==0.5.1

# API
fastapi==0.115.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0

uvicorn==0.30.6

# Interface
streamlit

# MLOps / tracking
mlflow==2.17.2

# Tests
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2

# Notebook
jupyter==1.0.0
ipykernel==6.29.3

torch==2.10.0
torchvision==0.25.0

python-multipart==0.0.9
